Move LLVMCompatibleBasicBlockBuilder to LLVM plugin
