+++ plugins/llvm/src/main/java/org/qbicc/plugin/llvm/LLVMCompatibleBasicBlockBuilder.java
@@ -1,4 +1,4 @@
-package org.qbicc.plugin.conversion;
+package org.qbicc.plugin.llvm;
 
 import java.util.List;
